requirements.txt
Flask==2.0.3
gunicorn==20.1.0
imageio==2.16.1
Jinja2==3.1.1
keras==2.8.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.0.1
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
sklearn==0.0
starlette==0.17.1
tensorboard==2.8.0
tensorflow-gpu==2.8.0
urllib3==1.26.9
utils==1.0.1
Werkzeug==2.0.3